Despliegue de CRUD completo

+++ backend/src/services/patient.service.ts
@@ -1,16 +1,6 @@
 import { type Patient } from '../interfaces/Patient'
 import PatientModel from '../models/Patient'
-
-const create = async (patientData: Patient): Promise<Patient> => {
-  try {
-    const patient = await PatientModel.create(patientData)
-    if (patient === null) throw new Error('Error al crear el paciente')
-    return patient
-  } catch (e) {
-    const error: string = e as string
-    throw new Error(error)
-  }
-}
+import UserModel from '../models/User'
 
 const getAll = async (): Promise<Patient[]> => {
   try {
@@ -24,7 +14,9 @@ const getAll = async (): Promise<Patient[]> => {
 
 const get = async (id: string): Promise<Patient> => {
   try {
-    const patient = await PatientModel.findById(id)
+    const user = await UserModel.findById(id)
+    if (user === null) throw new Error('Usuario no encontrado')
+    const patient = await PatientModel.findOne({ user: user._id })
     if (patient === null) throw new Error('Paciente no encontrado')
     return patient
   } catch (e) {
@@ -33,9 +25,11 @@ const get = async (id: string): Promise<Patient> => {
   }
 }
 
-const update = async (id: string, patientData: Patient): Promise<Patient> => {
+const update = async (id: string, patientData: any): Promise<Patient> => {
   try {
-    const patient = await PatientModel.findById(id)
+    const user = await UserModel.findById(id)
+    if (user === null) throw new Error('Usuario no encontrado')
+    const patient = await PatientModel.findOne({ user: user._id })
     if (patient === null) throw new Error('Paciente no encontrado')
     patient.birthdate = patientData.birthdate
     patient.phone = patientData.phone
@@ -50,8 +44,11 @@ const update = async (id: string, patientData: Patient): Promise<Patient> => {
 
 const deleteOne = async (id: string): Promise<Patient> => {
   try {
-    const patient = await PatientModel.findById(id)
+    const user = await UserModel.findById(id)
+    if (user === null) throw new Error('Usuario no encontrado')
+    const patient = await PatientModel.findOne({ user: user._id })
     if (patient === null) throw new Error('Paciente no encontrado')
+    await user.deleteOne()
     return await patient.deleteOne()
   } catch (e) {
     const error: string = e as string
@@ -59,4 +56,4 @@ const deleteOne = async (id: string): Promise<Patient> => {
   }
 }
 
-export { create, getAll, get, update, deleteOne }
+export { getAll, get, update, deleteOne }
